+++ src/main/java/com/logisticproject/services/cargoSortingLogics/cargoSortingMethods/ContainerCreationService.java
@@ -5,15 +5,15 @@ import org.springframework.stereotype.Component;
 @Component
 public class ContainerCreationService {
 
-    public Integer[][] create(int length, int width) {
-        Integer[][] array = new Integer[length + 2][width + 2];
+    public Double[][] create(int length, int width) {
+        Double[][] array = new Double[length + 2][width + 2];
         for (int i = 0; i < length + 1; i++){
             for (int j = 0; j <= width + 1; j++){
-                array[length + 1][j] = 0;
-                array[0][j] = 0;
+                array[length + 1][j] = (double) 0;
+                array[0][j] = (double) 0;
             }
-            array[i][width + 1] = 0;
-            array[i + 1][0] = 0;
+            array[i][width + 1] = (double) 0;
+            array[i + 1][0] = (double) 0;
         }
         return array;
     }
